feat(diag): make BlockProfiler a drop-in for old published images
Compile-portable across the plugin-API breaks between releases so the profiler
can be built against any published image's own assemblies and dropped into its
plugins/ dir without rebuilding the client.

- BlockProfilerPlugin: cast Block.Number to ulong at the target-set checks; the
  property returned long in older images (long->ulong is a binary break), so the
  cast keeps one source compilable against every generation.
- csproj: CopyLocalLockFileAssemblies so `dotnet publish` emits the JetBrains
  dependency chain next to the plugin dll.
- scripts/build-blockprofiler-dropin.sh: extract a target image's assemblies,
  reference-build the plugin against them, and emit the 4-file drop set.

Verified end-to-end (load + dep resolution + arm + collect, zero MissingMethod)
against nethermind/nethermind:1.36.0 and a 2-month-old master image.

+++ src/Nethermind/Nethermind.BlockProfiler/BlockProfilerPlugin.cs
@@ -81,7 +81,8 @@ public sealed class ProfilingBranchProcessor : IBranchProcessor, IDisposable
 
     private void OnBlockProcessing(object? sender, BlockEventArgs e)
     {
-        if (!_targets.Contains(e.Block.Number)) return;
+        // (ulong) cast keeps this source compilable as a drop-in against older images where Block.Number was long
+        if (!_targets.Contains((ulong)e.Block.Number)) return;
         if (_collecting) MeasureProfiler.StopCollectingData(); // close any stray window (e.g. a prior target that errored)
         MeasureProfiler.StartCollectingData();
         _collecting = true;
@@ -90,7 +91,7 @@ public sealed class ProfilingBranchProcessor : IBranchProcessor, IDisposable
 
     private void OnBlockProcessed(object? sender, BlockProcessedEventArgs e)
     {
-        if (!_collecting || !_targets.Contains(e.Block.Number)) return;
+        if (!_collecting || !_targets.Contains((ulong)e.Block.Number)) return;
         _collecting = false;
         MeasureProfiler.StopCollectingData();
         MeasureProfiler.SaveData(); // one snapshot per target block
